refactor: update Pots DTOs and service layer to use BigDecimal for target and current amounts, add pot update functionality

+++ src/test/java/fun/trackmoney/testutils/CreatePotsDTOFactory.java
@@ -3,12 +3,14 @@ package fun.trackmoney.testutils;
 import fun.trackmoney.pots.dtos.CreatePotsDTO;
 import fun.trackmoney.pots.enums.ColorPick;
 
+import java.math.BigDecimal;
+
 public class CreatePotsDTOFactory {
 
     public static CreatePotsDTO defaultCreatePot() {
         return new CreatePotsDTO(
             "Meu Cofrinho",
-            1000L,
+            BigDecimal.valueOf(1000),
             ColorPick.DARK_BLUE
         );
     }
@@ -16,14 +18,14 @@ public class CreatePotsDTOFactory {
     public static CreatePotsDTO vacationPot() {
         return new CreatePotsDTO(
             "Férias",
-            5000L,
+            BigDecimal.valueOf(5000),
             ColorPick.DARK_BLUE
         );
     }
 
     public static CreatePotsDTO customCreatePot(
             String name,
-            Long targetAmount,
+            BigDecimal targetAmount,
             ColorPick color
     ) {
         return new CreatePotsDTO(
